third_party/blink/web_tests/fast/encoding/css-charset-default-expected.txt

Test for bug 11011: External CSS is parsed as iso-8859-1 even though the main document is utf-8.

Stylesheet 1 (inherit document charset using xml-stylesheet processing instruction): "SUССЕSS"

Stylesheet 2 (inherit document charset using link): "SUССЕSS"

Stylesheet 3 (inherit referring stylesheet charset): "SUССЕSS"

Stylesheet 4 (inherit referring inline stylesheet charset): "SUССЕSS"
